Updated TechTip Model logging and notifications

+++ src/database/migrations/2024_04_06_002555_add_file_size_to_file_uploads.php
@@ -6,21 +6,20 @@ use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 use Illuminate\Support\Facades\Storage;
 
-return new class extends Migration
-{
+return new class extends Migration {
     /**
      * Run the migrations.
      */
     public function up(): void
     {
         Schema::table('file_uploads', function (Blueprint $table) {
-            $table->integer('file_size')->after('file_name');
+            $table->bigInteger('file_size')->after('file_name');
         });
 
         // Add the file size of each current file
         $allFiles = FileUpload::all();
         foreach ($allFiles as $fileData) {
-            $path = $fileData->folder.DIRECTORY_SEPARATOR.$fileData->file_name;
+            $path = $fileData->folder . DIRECTORY_SEPARATOR . $fileData->file_name;
             $size = -1;
 
             if (Storage::disk($fileData->disk)->exists($path)) {
